requirements.txt
git+git://github.com/Chia-Network/chia-blockchain.git@main#egg=chia-blockchain
blspy~=1.0.2
setuptools~=56.1.0
aiosqlite~=0.17.0
aiohttp~=3.7.4
pytest==6.2.4
pymongo==3.11
python-telegram-bot==v20.2